Release/v0.2.2 mlem

+++ api/src/main/java/io/pzstorm/storm/event/lua/OnObjectLeftMouseButtonDownEvent.java
@@ -36,14 +36,14 @@ public class OnObjectLeftMouseButtonDownEvent implements LuaEvent {
 	/**
 	 * Position of mouse along x-axis.
 	 */
-	public final Integer x;
+	public final Double x;
 
 	/**
 	 * Position of mouse along y-axis.
 	 */
-	public final Integer y;
+	public final Double y;
 
-	public OnObjectLeftMouseButtonDownEvent(IsoObject object, Integer x, Integer y) {
+	public OnObjectLeftMouseButtonDownEvent(IsoObject object, Double x, Double y) {
 
 		this.object = object;
 		this.x = x;
