+++ api/src/main/java/io/pzstorm/storm/event/lua/OnObjectRightMouseButtonUpEvent.java
@@ -36,14 +36,14 @@ public class OnObjectRightMouseButtonUpEvent implements LuaEvent {
 	/**
 	 * Position of mouse along x-axis.
 	 */
-	public final Integer x;
+	public final Double x;
 
 	/**
 	 * Position of mouse along y-axis.
 	 */
-	public final Integer y;
+	public final Double y;
 
-	public OnObjectRightMouseButtonUpEvent(IsoObject object, Integer x, Integer y) {
+	public OnObjectRightMouseButtonUpEvent(IsoObject object, Double x, Double y) {
 
 		this.object = object;
 		this.x = x;
